# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "qwen code API server",
    "description": "Building a blazing fast REST API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/sessions/create": {
      "post": {
        "tags": [
          "sessions"
        ],
        "description": "创建新的会话",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cwd": {
                    "type": "string"
                  },
                  "argv": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "cwd",
                  "argv"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessionId": {
                      "type": "string",
                      "description": "会话ID"
                    },
                    "session": {
                      "type": "object",
                      "description": "会话信息",
                      "properties": {
                        "sessionStartTime": {
                          "type": "string",
                          "format": "date-time",
                          "description": "会话开始时间"
                        },
                        "promptCount": {
                          "type": "number",
                          "description": "提示词计数"
                        },
                        "lastPromptTokenCount": {
                          "type": "number",
                          "description": "最后提示词token计数"
                        },
                        "metrics": {
                          "type": "object",
                          "description": "指标数据",
                          "properties": {
                            "models": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "object",
                                "title": "ModelMetrics",
                                "properties": {
                                  "api": {
                                    "type": "object",
                                    "properties": {
                                      "totalRequests": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总API请求次数"
                                      },
                                      "totalErrors": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总API错误次数"
                                      },
                                      "totalLatencyMs": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总延迟时间（毫秒）"
                                      }
                                    },
                                    "required": [
                                      "totalRequests",
                                      "totalErrors",
                                      "totalLatencyMs"
                                    ],
                                    "additionalProperties": false
                                  },
                                  "tokens": {
                                    "type": "object",
                                    "properties": {
                                      "prompt": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "提示词token数量"
                                      },
                                      "candidates": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "候选响应token数量"
                                      },
                                      "total": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总token数量"
                                      },
                                      "cached": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "缓存token数量"
                                      },
                                      "thoughts": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "思考过程token数量"
                                      },
                                      "tool": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "工具调用token数量"
                                      }
                                    },
                                    "required": [
                                      "prompt",
                                      "candidates",
                                      "total",
                                      "cached",
                                      "thoughts",
                                      "tool"
                                    ],
                                    "additionalProperties": false
                                  }
                                },
                                "required": [
                                  "api",
                                  "tokens"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "tools": {
                              "type": "object",
                              "description": "工具调用统计",
                              "properties": {
                                "totalCalls": {
                                  "type": "number",
                                  "description": "总调用次数"
                                },
                                "totalSuccess": {
                                  "type": "number",
                                  "description": "成功调用次数"
                                },
                                "totalFail": {
                                  "type": "number",
                                  "description": "失败调用次数"
                                },
                                "totalDurationMs": {
                                  "type": "number",
                                  "description": "总耗时(毫秒)"
                                },
                                "totalDecisions": {
                                  "type": "object",
                                  "description": "决策统计",
                                  "properties": {
                                    "accept": {
                                      "type": "number",
                                      "description": "接受次数"
                                    },
                                    "reject": {
                                      "type": "number",
                                      "description": "拒绝次数"
                                    },
                                    "modify": {
                                      "type": "number",
                                      "description": "修改次数"
                                    }
                                  }
                                },
                                "byName": {
                                  "type": "object",
                                  "description": "按工具名称统计",
                                  "properties": {}
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessionId": {
                      "type": "string",
                      "description": "会话ID"
                    },
                    "session": {
                      "type": "object",
                      "description": "会话信息",
                      "properties": {
                        "sessionStartTime": {
                          "type": "string",
                          "format": "date-time",
                          "description": "会话开始时间"
                        },
                        "promptCount": {
                          "type": "number",
                          "description": "提示词计数"
                        },
                        "lastPromptTokenCount": {
                          "type": "number",
                          "description": "最后提示词token计数"
                        },
                        "metrics": {
                          "type": "object",
                          "description": "指标数据",
                          "properties": {
                            "models": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "object",
                                "title": "ModelMetrics",
                                "properties": {
                                  "api": {
                                    "type": "object",
                                    "properties": {
                                      "totalRequests": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总API请求次数"
                                      },
                                      "totalErrors": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总API错误次数"
                                      },
                                      "totalLatencyMs": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总延迟时间（毫秒）"
                                      }
                                    },
                                    "required": [
                                      "totalRequests",
                                      "totalErrors",
                                      "totalLatencyMs"
                                    ],
                                    "additionalProperties": false
                                  },
                                  "tokens": {
                                    "type": "object",
                                    "properties": {
                                      "prompt": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "提示词token数量"
                                      },
                                      "candidates": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "候选响应token数量"
                                      },
                                      "total": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "总token数量"
                                      },
                                      "cached": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "缓存token数量"
                                      },
                                      "thoughts": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "思考过程token数量"
                                      },
                                      "tool": {
                                        "type": "number",
                                        "minimum": 0,
                                        "description": "工具调用token数量"
                                      }
                                    },
                                    "required": [
                                      "prompt",
                                      "candidates",
                                      "total",
                                      "cached",
                                      "thoughts",
                                      "tool"
                                    ],
                                    "additionalProperties": false
                                  }
                                },
                                "required": [
                                  "api",
                                  "tokens"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "tools": {
                              "type": "object",
                              "description": "工具调用统计",
                              "properties": {
                                "totalCalls": {
                                  "type": "number",
                                  "description": "总调用次数"
                                },
                                "totalSuccess": {
                                  "type": "number",
                                  "description": "成功调用次数"
                                },
                                "totalFail": {
                                  "type": "number",
                                  "description": "失败调用次数"
                                },
                                "totalDurationMs": {
                                  "type": "number",
                                  "description": "总耗时(毫秒)"
                                },
                                "totalDecisions": {
                                  "type": "object",
                                  "description": "决策统计",
                                  "properties": {
                                    "accept": {
                                      "type": "number",
                                      "description": "接受次数"
                                    },
                                    "reject": {
                                      "type": "number",
                                      "description": "拒绝次数"
                                    },
                                    "modify": {
                                      "type": "number",
                                      "description": "修改次数"
                                    }
                                  }
                                },
                                "byName": {
                                  "type": "object",
                                  "description": "按工具名称统计",
                                  "properties": {}
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/list": {
      "get": {
        "tags": [
          "sessions"
        ],
        "description": "获取所有会话ID列表",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "会话ID列表"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "会话ID列表"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/cwd": {
      "post": {
        "tags": [
          "sessions"
        ],
        "description": "根据工作目录路径查找会话",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cwd": {
                    "type": "string",
                    "description": "要查找的工作目录路径"
                  }
                },
                "required": [
                  "cwd"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "会话ID列表"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "是否成功"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "message": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "会话ID列表"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}